move element code up in the tree
